Move ScopeValidatorInterface to PromotionBundle
